backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bump-pydantic==0.8.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.100.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
libcst==1.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.6.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.9.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.2
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
